chore: move hook to hooks folder
